SystemCommandStore moved to an interface implemented by DefaultSystemCommandStore.
